Fix error: validator not defined
For some reason the static rules function is being called on a form that the validator was never bound to. This extra check will at least suppress this error until someone can figure out why this function is being called referencing a form that was never associated with the validator.