Remove focusin/focusout polyfill
Removed polyfill for focusin and focusout events as this implementation is problematic in IE9, and jQuery core has handled them since 1.4 (1.8.2 is the latest at this point in time). Note that this raises the minimum compatibility of the validation plugin to jQuery 1.4, as per https://github.com/jzaefferer/jquery-validation/issues/542.