min/max validation for text interpreted as number
Issue #639 notes that min/max validation for input type text
is no longer interpreted numerically in 1.11.0.  This differs
from 1.10.0, where it was interpreted numerically, even though
html5 spec says the input type text cannot have min/max attributes.

This issue was caused when adding support for dates, where
interpreting min/max as number in 1.10.0 caused the error
that "2012-11-22" converted to number is NaN.

This commit once more interprets min/max for text numerically;
input type=date remains non-numerical.

It builds on a proposed fix for #648, that interprets
input with type number/range min/max numerically.

The effect is:

 * input type=text has numeric min/max, as in 1.10.0
 * input type=date has working min/max for type date;
   on mobile browsers you also get a date picker,
   plus the browser may reject invalid dates before
   javascript gets a chance to complain.
 * input type=number or range get numeric min/max,
   plus numeric keypad or slider on mobile browsers,
   plus browser may reject invalid input before javascript
   gets a chance to complain

Allowing use of min/max with type=number/range/date is important
for mobile browsers, where the numeric keypad or date picker
make the input much easier to use than a generic text input field.
In this situation jquery-validate remains necessary to support
older browsers that do not do input validation based on type
and min/max.

For situations where numeric input should be validated by jquery
without giving the browser a chance to validate the input format,
input type=text in combination with min/max can be used, as in 1.10.0.

See also #648.

Validation of this commit is problematic.  It passes validation
in the Chromium browser, and looks OK in a simple sample page
such as http://jsbin.com/awokex/3, but fails in grunt with
the message "Please enter a valid date."

I suspect this is caused by a bug in phantomjs:
http://code.google.com/p/phantomjs/issues/detail?id=187#c6

A fix for this bug exists:
https://github.com/ariya/phantomjs/commit/35c1971595
but that is released only in phantomjs 1.8.0, and the version
that I see grunt use is only 1.7.0.  Sigh,