Problem with not running validation when onfocusout is called.
There is a problem with plugin not validating fields when focus is out. This was reported by some people to work on "random" basis.

The problem is on line 239. The check fails because `this.optional` returns string `"dependency-mismatch"` and the whole if statement returns wrong bool result. Sometimes it works, because the field is in `this.submited` array

The whole method `optional` should be probably changed.
This change fix problem for me, but I'm not sure about other use cases.

I'm not knee deep in the code.