Ability to specify message parameters separately from rule parameters
I find it very restrictive that the rule's parameters are used for the
message and there is no way to specify overt message parameters. There
are many custom rules that need to specify different message parameters
than rule parameters. For example, when validating that one field must
be less than other, the rule parameter needs to be the selector of the
other field, but the message is "{0} must be less than {1}" so the
message parameters need to be something like "The minimum" and "the
maximum."