Support for class-based messages
I figure if you can specify a class-based rule, it makes sense to also
allow a custom message for that class-based rule by class, just as you
can for a custom static rule. For example, if I want to override the
custom message for a rule, but I need it to be class-based so I can
clone it with jQuery, there was no way of doing that.