Core: Update elementValue method to deal with type="number" fields
The HTML5 draft defines, for type="number" inputs: "The value
sanitization algorithm is as follows: If the value of the element is not
a valid floating-point number, then set it to the empty string instead."
If the constraint validation considers the input invalid, return false
as the value, instead of the empty string, for the number and digit
methods to output their messages.

This would break in browsers that support type="number", but not the
constraint validation API. I don't know of any existing browser where
that applies.

Fixes #858
Fixes #922
Closes #1093