Allow focusout event for hidden elements
Allow focusout event for hidden elements. If you have a custom styled form element (ie a dropdown with ul/li structure and a hidden element for the value), and you want to apply form errors on blur, calling blur on the hidden element doesn't work. This fix will allow a blur event on a hidden element.