Update README.md
The link to the description of the email validation in the HTML5 spec is broken. Looking at the link in the wayback machine (http://web.archive.org/web/20141020123432/https://html.spec.whatwg.org/multipage/forms.html#e-mail-state-(type=email)) it looks like it previously redirected to a page similar to the one I've suggested.