Localization: Remove full stop at the end of italian messages
It looks ugly to see a full stop at the end of an error message, and it
is worse if the message is used in dynamic context (e.g. in a kind of
popup message like a tooltip)

Also use the default (possibly localized) error message when a wrong date
is used instead of using a static message to improve flexibility, for all
three custom date methods.

Closes #1326