Add stopPropagation setting
When integrating with other JS libraries, such as those that do Ajax requests on form submit, we want to make sure that the form does not actually get submitted when it is invalid. We do this by calling ``stopImmediatePropagation()`` on the event when this setting is true. This stops all other listeners from getting the submit event and therefore the form is not submitted.