min rule
If the filed is allowed to be number then it accepts comma in value (i.e. 1200 and 1,200 both are valid). Now if "min" rule is provided with number then currently even if i'm entering 1,200 error(Please enter a value greater than or equal to 0.) is throwing. So if the current element has number and min both rules then we need to replace comma with 0 length string.