collect raised error in an object fixes ##1925
All errors that are raised will be collected into a raisedErrors object on the base validator object (which can be accessed from an).  The id of the element will be used as the key with an empty object as the value. Each corresponding error with have the method name (such as "required") as the key with the error message as the value.

So for instance... if you have the following rules and messages:

rules: {
    email: { required: true, email: true }
},
messages: {
    email: {required: "Enter your email address", email: "A valid email address is required"}
},

would have a rasiedErrors object of:

this.rasiedErrors["email"]["required"] = "Enter your email address"
this.rasiedErrors["email"]["email"] = "A valid email address is required"