Core: guard against null & undefined values in required method
An input element will never have undefined or null as value
but the normalizer may return null or undefined for an element.

Also, jQuery (< 3.x) returns `null` for select element with the
multiple attribute set if no options are selected.

As a result, the required method should guard against that.